Add new test-lib functions generalizing test_cmp_rev to several revisions
The new function test_revs_have_expected is for the case that we expect
one given revision in a list of actual revisions.

The new function test_rev_among_expected is for the case that we have
a list of expected revisions and want to check if the actual revision
is contained in that list.