Improve test_cmp_rev() in test-lib-functions.sh
test_cmp_rev() has the (undocumented) side effects that it generates files
expect.rev and actual.rev in the repository.
This makes it hard to use this auxiliary functions when one wants to
maintain a clean worktree in the repository.

This commit changes test_cmp_rev() such that it does not generate any
files. The readability of the error output is improved and it is
made clear what the actual and what the expected argument is.

There is no need to change any tests because the presence of the files
expect.rev and actual.rev is not used by any tests.